issue #118 - @Ocramius review
